Minor improv (#487)
* feat: always display product description if available

This commit modifies the ProductDetail component to always display the product description if it exists, regardless of whether an HTML file is present. This ensures that users always see the product description when it is available.

* feat: enhance product download with metadata and remove Github login

This commit introduces enhancements to the product download functionality and removes the Github login option from the frontend.

- **Backend:**
  - Adds a `download` action to the `ProductViewSet` that generates a YAML file containing product metadata, including associated columns, and includes it in the downloaded ZIP archive.
  - The metadata file (`product_metadata.yaml`) provides comprehensive information about the product, such as file details (name, type, extension, size, number of rows), associated columns (column name, UCD, alias), and delimiter (for CSV files).
  - Refactors the `__get_product_contents` and `__get_full_product` methods to retrieve and format product content and file information.

- **Frontend:**
  - Removes the Github login button and related styling from the login page.
  - Removes the `GithubUrl` prop from the `Login` component and its usage.

* fix: remove release path validation

This commit removes the validation check that ensures the release path exists as a directory or symlink.

The validation was causing issues in some deployment scenarios where the release path might not be immediately available when the process view is accessed. Removing the check allows the process to proceed even if the release path is not initially present, assuming it will become available later in the process.