feat: remove links from TrainingSetMaker and SpeczCatalogs
This commit removes the Link component and adds a "(soon)" suffix to the titles of the TrainingSetMaker and SpeczCatalogs components.

The Link component was likely used to navigate to the respective pages, but has been removed in this update. The addition of "(soon)" indicates that these features are not yet available.